requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
requests==2.32.3
requests-cache==1.2.1
aiohttp==3.10.5
beautifulsoup4==4.12.3
yfinance==1.2.0
pandas==2.2.2
plotly==5.24.1
pydantic==2.9.2
python-dotenv==1.0.1
